# Optimize In performance by reducing allocations for common queries