```
name          old time/op    new time/op    delta
In-8             466ns ± 1%     328ns ± 0%  -29.69%  (p=0.000 n=10+10)
In1k-8          6.42µs ± 0%    7.49µs ± 0%  +16.73%  (p=0.000 n=8+9)
In1kInt-8       9.71µs ± 8%    8.32µs ± 0%  -14.38%  (p=0.000 n=10+10)
In1kString-8    9.86µs ± 1%    9.05µs ± 0%   -8.22%  (p=0.000 n=10+9)

name          old alloc/op   new alloc/op   delta
In-8              512B ± 0%      272B ± 0%  -46.88%  (p=0.000 n=10+10)
In1k-8          22.7kB ± 0%    19.5kB ± 0%  -14.16%  (p=0.000 n=10+10)
In1kInt-8       22.7kB ± 0%    19.5kB ± 0%  -14.16%  (p=0.000 n=10+10)
In1kString-8    22.7kB ± 0%    19.5kB ± 0%  -14.16%  (p=0.000 n=10+10)

name          old allocs/op  new allocs/op  delta
In-8              10.0 ± 0%       4.0 ± 0%  -60.00%  (p=0.000 n=10+10)
In1k-8            5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.000 n=10+10)
In1kInt-8         5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.000 n=10+10)
In1kString-8      5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.000 n=10+10)
```